(BSR)[API] refactor: Re-use `check_user_has_access_to_offerer()`
Don't define another function that does the same thing.

+++ api/src/pcapi/routes/pro/venue_providers.py
@@ -1,6 +1,8 @@
 from flask_login import current_user
 from flask_login import login_required
+from werkzeug.exceptions import NotFound
 
+import pcapi.core.offerers.models as offerers_models
 from pcapi.core.providers import api
 from pcapi.core.providers import exceptions
 from pcapi.core.providers import repository
@@ -14,17 +16,25 @@ from pcapi.routes.serialization.venue_provider_serialize import ListVenueProvide
 from pcapi.routes.serialization.venue_provider_serialize import PostVenueProviderBody
 from pcapi.routes.serialization.venue_provider_serialize import VenueProviderResponse
 from pcapi.serialization.decorator import spectree_serialize
-from pcapi.validation.routes.users_authorizations import check_user_can_alter_venue
+from pcapi.utils import rest
 from pcapi.workers.venue_provider_job import venue_provider_job
 
 from . import blueprint
+
+
+def _get_venue_or_404(venue_id: int) -> offerers_models.Venue:
+    venue = offerers_models.Venue.query.filter_by(id=venue_id).one_or_none()
+    if not venue:
+        raise NotFound
+    return venue
 
 
 @private_api.route("/venueProviders", methods=["GET"])
 @login_required
 @spectree_serialize(on_success_status=200, response_model=ListVenueProviderResponse, api=blueprint.pro_private_schema)
 def list_venue_providers(query: ListVenueProviderQuery) -> ListVenueProviderResponse:
-    check_user_can_alter_venue(current_user, query.venue_id)
+    venue = _get_venue_or_404(query.venue_id)
+    rest.check_user_has_access_to_offerer(current_user, venue.managingOffererId)
 
     venue_provider_list = repository.get_venue_provider_list(query.venue_id)
     for venue_provider in venue_provider_list:
@@ -40,7 +50,8 @@ def list_venue_providers(query: ListVenueProviderQuery) -> ListVenueProviderResp
 @spectree_serialize(on_success_status=201, response_model=VenueProviderResponse, api=blueprint.pro_private_schema)
 def create_venue_provider(body: PostVenueProviderBody) -> VenueProviderResponse:
     body.venueIdAtOfferProvider = None
-    check_user_can_alter_venue(current_user, body.venueId)
+    venue = _get_venue_or_404(body.venueId)
+    rest.check_user_has_access_to_offerer(current_user, venue.managingOffererId)
 
     try:
         new_venue_provider = api.create_venue_provider(
@@ -108,7 +119,8 @@ def create_venue_provider(body: PostVenueProviderBody) -> VenueProviderResponse:
 @login_required
 @spectree_serialize(on_success_status=200, response_model=VenueProviderResponse, api=blueprint.pro_private_schema)
 def update_venue_provider(body: PostVenueProviderBody) -> VenueProviderResponse:
-    check_user_can_alter_venue(current_user, body.venueId)
+    venue = _get_venue_or_404(body.venueId)
+    rest.check_user_has_access_to_offerer(current_user, venue.managingOffererId)
 
     venue_provider = get_venue_provider_by_venue_and_provider_ids(body.venueId, body.providerId)
 
@@ -124,8 +136,9 @@ def update_venue_provider(body: PostVenueProviderBody) -> VenueProviderResponse:
 @spectree_serialize(on_success_status=204, api=blueprint.pro_private_schema)
 def delete_venue_provider(venue_provider_id: int) -> None:
     venue_provider = repository.get_venue_provider_by_id(venue_provider_id)
-
-    check_user_can_alter_venue(current_user, venue_provider.venueId)
+    if not venue_provider:
+        raise NotFound()
+    rest.check_user_has_access_to_offerer(current_user, venue_provider.venue.managingOffererId)
 
     api.delete_venue_provider(venue_provider, author=current_user)
